修正指摘(1回目)

+++ app/controllers/items_controller.rb
@@ -37,9 +37,12 @@ class ItemsController < ApplicationController
   end 
 
   def destroy
-    if @item.destroy
-     redirect_to root_path
+    if current_user.id != @item.user_id
+      redirect_to root_path
     end
+    if @item.destroy
+    end
+    redirect_to root_path
   end
 
   private
